bunch of lint, and commandline edit rework

+++ b0_exec.h
@@ -1,10 +1,11 @@
 #ifndef _EXEC_H
 #define _EXEC_H 1
 
+#include "types.h"
 #include "banking.h"
 
-int exec(char* command);
+int16_t exec(char* command);
 
-DECLARE_BANKED(exec, BANK(0), int, bk_exec, (char* command), (command))
+DECLARE_BANKED(exec, BANK(0), int16_t, bk_exec, (char* command), (command))
 
 #endif
